multi game

+++ app/src/modules/game.py
@@ -182,7 +182,8 @@ async def hanler_choise_game(chat_id, category, jawab = False):
         id_msg_jwb = send_msg_jawab.id
         if in_game_chat_id[chat_id]["id_msg_jwb"]:
             id_msg = in_game_chat_id[chat_id].get('id_msg_jwb', None)
-            await bot.delete_messages(chat_id, id_msg)
+            if id_msg:
+                await bot.delete_messages(chat_id, id_msg)
             in_game_chat_id[chat_id].update({"id_msg_jwb": id_msg_jwb})
         else:
             in_game_chat_id[chat_id]["id_msg_jwb"] = {"id_msg_jwb": id_msg_jwb}
@@ -484,8 +485,9 @@ async def check_answer(client, message: Message):
                         # format_jawab = f"Jawaban {mention} benar!\n tunggu 5 detik untuk next soalll.."
                         format_jawab = f"Jawaban {mention} benar!\n"
                         jawab = await bot.send_message(chat_id, format_jawab, protect_content=True)
-                        await hanler_choise_game(chat_id, category, jawab=True)
+                        await asyncio.sleep(2)
                         await bot.delete_messages(chat_id, jawab.id)
+                        await hanler_choise_game(chat_id, category, jawab=True)
                 else:
                     jawaban = question["jawaban"]
                     jawab_user = message.text.strip().lower()
